docs: use our docs url for the ogp:url properties Sphinx generates (#1411)
Currently we use the Open Graph configuration
from the canonical-starter-pack without any modifications, but we should
use the hosting url for our docs instead of theirs.

The Open Graph configuration is used by Sphinx to generate the Open
Graph Protocol (ogp) properties in our docs pages. This information is
used when generating link previews, for example when linking to our docs
in Mattermost messages.

In practice these `ogp:url`s seem to be ignored in link previews, but let's
fix them anyway since it's simple.