docs: the very start of a team Python style guide (#1409)
This adds the very start of a team Python style guide, with a few items
we've discussed recently. The idea is to add as things come up (more
than once) in PRs and we feel it's worth adding them to the style guide
to avoid having to make a specific decision each time.