ci: disable the automatic ops-scenario releasing (#1415)
This hasn't been fully updated to handle the location in the merged
repo, but has been enough to wrongly release an `ops.version.version`
release (and tag) on merge to main.

Disable this for now so that we don't accidentally have another
`ops.version.version` release. I'll do a follow-up PR that properly
implements ops-scenario releasing alongside ops releases.