Properly use _JujuContext.
Create one single _JujuContext object in _Ops (which will later be a _Manager instance) and pass it around, rather than creating new ones all the time.

A bit awkward in a couple of places because some of the mechanisms are technically public methods so ought to keep backwards compatibility.

This doesn't extend the use in any way yet - for example, when we are setting up the Juju environment variables based on the event.