fix: make push_path open in binary mode so it works on non-text files (#1458)
There was a bug in `push_path` where it called `open` with the default
(text) mode on the local files, causing it to not work on binary files
(more specifically, if they weren't valid UTF-8). This is the fix for
that -- `push_path` should treat the files as raw bytes, so read in
binary mode.

Also change the test to ensure binary / non-UTF-8 files work as
expected.

Fixes #1455.