chore: minor linting improvements to testing/src/scenario (#1456)
We're not aiming to get 100% in the TIOBE scoring, but it's nice to get
a higher score when it aligns with our existing style and doesn't
involve significant changes. The `testing` folder already scores 95.99
but there's still some low-hanging fruit, which this PR addresses:

* Some basic import reordering, which is mostly because "ops" is
first-party now.
* Clearing out outdated TODO/FIXME style comments (I've considered each
of these, and the ones removed either are out of date, are not something
we want to do now, or already have tickets).
* TIOBE's linter prefers multiple `if` statements rather than
if/elif/else when the block unconditionally returns/raises. This is a
debatable style choice, but where it seemed reasonable I've applied that
here. I don't think we have a team style decision on this yet.
* Added module, class, method docstrings.