feat: Make Context and Manager variadic types (#1445)
Closes #1444.

This change enables passing the Charm type all the way from the Context
to the charm attribute in the manager for better autocompletion and type
checks.

You can give this a try with the following code snippet:

```python
from ops import CharmBase
from testing.src.scenario import Context, State


class MyCharm(CharmBase):
    some_attribute: str


def test_function():
    ctx = Context(MyCharm)
    state = State()

    with ctx(ctx.on.config_changed(), state) as manager:
        charm = manager.charm
        charm.some_attribute  # behold, autocompletion!
```

A word of caution, though, the testing module is not using the bundled
scenario code base. It is still using `ops_scenario`. For users to
benefit from this new feature, we would need to change the import
machinery.