fix: require the same object to be in the testing state as in the event (#1468)
This was incorrectly changed during the import of the ops-scenario code
to this repo. As discussed in #1429, having an object in the state that
has the same identifier as the object in the event is insufficient and
could end up being confusing: it should be the same actual object.

As well as reverting the previous change, adjust the error messages so
that they're clearer if this case is triggered.

Fixes #1429