ci: don't cancel other tests when one fails
A test can pass on one python/os version and fail on others, and it’s
nice to see that in the outcome, rather than dig into cancelled tests
which may or may not have hit the failing case.

Since unit tests are cheap and fast, we’re not really saving much by
cancelling them – by the time one unit test run fails and asks the
others to cancel, they’ll already have been spun up and run many of
the tests, so canceling them doesn't save much.

The pip-install and test-real-pebble jobs are faster than the test job,
so this commit adds fail-fast: false to them too.

Note: test with macos-latest and python 3.10 is the slowest by far.