docs: use the right ops-scenario for building the docs (#1470)
This PR fixes the version of ops-scenario used when building the docs.

Previously, the published version of ops-scenario was used for
generating the documentation, and the generated requirements file was
used for local builds but not for read-the-docs.

That breaks if there are local incompatible changes (such as renaming a
module), and is generally wrong anyway.

With this PR, the local version is loaded, and read-the-docs uses the
same requirements file as building locally. A few minor type annotations
were adjusted to remove remaining warnings (I don't know why Sphinx
couldn't figure out where `Path` was coming from, but keeping it in the
`pathlib` namespace is generally nicer anyway).

Split out from #1467.