docs: go full Diátaxis, ingesting the relevant juju.is/docs/sdk documentation (#1481)
[**Preview on
ReadTheDocs**](https://ops--1481.org.readthedocs.build/en/1481/)

This PR adds all the juju.is/docs content featuring Ops.

The content was originally formatted for Discourse; this PR reformats it
to serve the RTD project the docs are destined for.

The PR also retitles all how-to guides as "Manage x", often with
sections of the form "Implement the feature" and "Test the feature".
This extends a pattern we already had in juju docs, which will continue
in the juju RTD docs, and which we've also implemented in the new
charmcraft RTD docs. Whereas in the past we tended to associate "manage"
with a juju admin, here it's used as a catch-all for "how-to do all the
things that it is possible to do about x in juju/charmcraft/ops". I
believe the pattern is no clumsier than what we had before and having
the same terminology across juju/charmcraft/ops , along with
cross-referencing, will make it easier for people to piece together the
story. (E.g., in Juju, "manage relations" will include `juju integrate`,
while in charmcraft it'll mean declaring a particular endpoint and in
ops it'll mean reacting to certain events and using or implementing an
interface.)

Finally, the PR trims down the Reference and Explanation docs to just
content directly relevant to ops. E.g., the charm taxonomy, events /
lifecycle bits, charm maturity, charm best practices are all things that
belong at a higher level of abstraction, directly in juju docs. E.g., in
the charm best practices doc the first best practice we speak about is
using ops. (A lot of that content should also be a lot leaner -- e.g.,
I'm working to incorporate all the charm SDK event content into the juju
doc on hooks, and the charm maturity and charm best practices could
really use a complete rethink as well, as they are way too verbose,
always understandable, and often duplicate content that's better
expressed at a more specific level.)

Some things that I think should be addressed in subsequent PRs:

- Everywhere: (1) The docs reference juju and charmcraft pages. As those
pages are not yet live, the links are empty. Once the pages are live,
which should be very soon, the links should be updated. Note: This does
create a temporary crisis in the ops RTD docs which, unlike the juju and
the charmcraft RTD docs, are actually already live. I believe that's
acceptable, especially since (a) at present people still only look at
the ops RTD docs for reference and (b) the crisis will be resolved very
soon. (2) The testing docs, from the tutorial through the how-to guides
and all the way to reference, should be thoroughly revisited to reflect
(a) the fact that scenario is now the de facto unit testing tool and
that harness is legacy and (b) the fact that most of our "how to manage
x" docs have sections on testing"; in short, I anticipate that a lot of
that stuff will end up being removed altogether, the result being much
leaner and more relevant docs.
- Tutorial: The Kubernetes tutorial should use `charmcraft init`
(instead of creating all the files by hand) and should only focus on
features that are aided by ops (i.e., the publish section should be made
part of the planned charmcraft tutorial featuring a regular,
non-12-factor app, charm). Also, it's a good idea to make it quite a bit
shorter. (When we first created it, it was trying to make up for the
fact that our how-to guides were not very helpful. At present, however,
I believe it can be much leaner.)
- How-to guides: The "Manage x" formula, especially as extended just now
in the new charmcraft RTD docs, where it includes a "Manage charms",
suggest we should perhaps have a "Manage charms" how-to in the ops docs
too. If in Juju managing charms means, e.g., `juju find`, `juju
download`, `juju deploy`, and `juju refresh`, and if in charmcraft it
means `charmcraft init`, ... `charmcraft pack`, `charmcraft upload`,
etc., in ops it might be the place where we give the recipe for the
general charm logic from the point of view of ops -- import ops,
logging, creating a charm class that inherits from CharmBase, the init
method, etc. PS I anticipate some of the content from "Run workloads
with..." and the logs how-to might find their home here.

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>
Co-authored-by: Tony Meyer <tony.meyer@canonical.com>
Co-authored-by: Dave Wilding <tech@dpw.me>
Co-authored-by: Tony Meyer <tony.meyer@gmail.com>