With the benchmark tests in a separate folder, no need to change pyproject.toml
Ideally, we would statically check these, like the other tests, but it does not seem super important for benchmarking tests. The issue is that there is an import after a sys.path manipulation, and pyright cannot figure that out, so complains about a lot of things. We do not want to have that charm on the path, and adjusting the path for pyright does not seem great either. Not clear what the cleanest solution is.