docs: fix Markdown syntax in ops.testing readme (#1502)
The previous syntax made Markdown treat the entire paragraph as the
title. I don't think this was intended. Maybe there was originally a
list or something else at this point, but this changes the formatting
back to a normal title for a single sentence, and then the following
paragraph is formatted normally.

Copied from @jameinel 's PR in the old repo:
https://github.com/canonical/ops-scenario/pull/205

(I accidentally made this change in
0c55e6473c604c7f041fbc92201d59429fec070d).