test: add a small set of ops.testing benchmark tests (#1504)
The PR:
 * Adds a new `tox -e benchmark` command.
 * Adds a new `test/test_benchmark.py` model.
 * Excludes the benchmark tests from `tox -e unit` and `tox -e coverage`

I wrote a small set of benchmark tests while working on #1434. It seems
like it'd be worth keeping these - in the future, they could be expanded
to include a small set of benchmark tests that target ops more
specifically, and ideally we could have a CI workflow that failed if
there were regressions (currently, running the tests just outputs the
timing, and you can use the pytest-benchmark tools to do comparisons).