fix: remove ops.main.main deprecation warning, and avoid warnings in action output (#1496)
Background:
* A `warnings.warn` call outputs to stderr by default.
* Juju captures stderr and when executing an action provides it to the
user.
* We emit a DeprecationWarning when charms call `main()` other than
`import ops ... ops.main()`
* DeprecationWarnings are ignored by default, *except* for issues in the
`__main__` module.
* A warning like this is meant to be discovered during tests - but unit
tests typically do not call `main()` or execute the `charm.py` file, so
they will not trigger this warning. That means that we're relying on
integration tests.

The above means that, currently, the deprecation warning is included in
the action output - this warning is intended for the developer of the
charm, not the user of the charm, so we want to change this.

Firstly, the `ops.main.main` deprecation warning is removed. It's too
awkward to have a warning that gets emitted before the framework is set
up. Instead, we'll rely on `charmcraft analyse` and static checks.

However, while that will solve the specific issue in #1460, we do also
want to avoid any future leakage of warnings into action output if there
are other warnings emitted that originate in the `__main__` model.

To that, rather than rely on Juju capturing the warning via stderr, we
use the logging package to capture all warnings and redirect them to
logging calls. These are then directed to Juju's debug-log via the
existing mechanism. While we are doing this, we minimise the output to
include references to the filename and line, but not the actual code
itself (which may be sensitive).

When running (Scenario) tests, we ensure that the warning and logging
systems use their default behaviour, so that pytest and other tools can
work as expected.

Fixes #1460