chore: update documentation dependencies (#1526)
There's a security release for Jinja (not one that impacts our use from
what I can tell, but nice to just move to the updated version anyway).
While bumping that, bump all the other dependencies as well.

This is just the result of removing the requirements file and
regenerating it with `tox -e docs-deps`.