chore: fix compatibility with ops-scenario 7.0.5 (#1543)
The next version of ops-scenario renames the `runtime` module to
`_runtime`. The `main` branch of `ops` currently expects that rename to
have happened, but doesn't require it in the `[testing]` dependency.
This means that `main@HEAD` `ops` doesn't work unless you also use
`main@HEAD` of `ops-scenario`, which isn't the case in our compatibility
checks - or in any charm that ends up with that combination.

We could fix this by adjusting the `ops` dependencies to require the
newer `ops-scenario`, but that would really require releasing a new
version of `ops-scenario`, and we're unlikely to be doing that before we
also release `ops`.

The relevant code is only required until we release the next version of
`ops-scenario` so for now just handle 7.0.5 as well as the newer code.

This will fix several of the "broad compatibility" charm CI tests.