ci: use Concierge to set up the smoke test environments (#1541)
The smoke tests are currently failing because of `sg` behaviour changes
in the `ubuntu:latest` runners. Rather than figure out what the solution
for that is, this PR simplifies the entire environment setup to just use
Concierge instead.

This also switches from testing against Juju 3.5 to 3/latest (which is
3.6), since 3.6 is available now and the intention was always to test
against the latest 3.

A [run can be seen in my
fork](https://github.com/tonyandrewmeyer/operator/actions/runs/13003299600).