fix: put the Juju version in the environment as well as the JujuContext (#1559)
Charms are expected to access the Juju environment variables via the
mechanisms that ops provides. Generally, this is through the model and
other attributes of the charm class. However, the Juju version is not
available via the charm object, and instead is expected to be obtained
by charms via `JujuVersion.from_environ()`.

Previously, we stopped Scenario from putting all the Juju environment
variables in `os.environ`, and instead provided them directly to the
`JujuContext` object. This works, except for the `JUJU_VERSION` case, as
`from_environ` (as the name indicates) loads the version from
`os.environ`, not from a `JujuContext` object.

This PR changes the behaviour so that `JUJU_VERSION` is set in both the
context and the environment.

Fixes #1555.