Skipping the double emit at the end is complex, so don't bother for now.
It's probably less of a shock to have them both, in case charms are relying on the event being last, and making the reemit aware of the current event or emit aware of what reemit did are both complex changes and it's not clear if it's really worth it. In many cases the deferred event and current event won't be the same anyway, so it won't matter.