fix: ensure that the event snapshot is available when one observer defers and the other does not (#1562)
If an event has multiple observers, then if *any* of the
(event+observer) pairs do not defer, then we need to save the snapshot
so that it's available to process that (event+observer).

#1372 introduced a bug where if *any* of the (event+observer) pairs was
skipped because an exact match was already in the deferral queue, we
would skip saving the snapshot. This is ok if there's only a single
observer, or if an (event+observer) that did not defer was processed
before any that did defer, but not otherwise.

Fixes #1561