docs: fill out remaining external link placeholders (#1564)
All the remaining "UPDATE LINKS" notes in the docs are replaced with
actual links to the external Juju/Charmcraft docs.

Also changes the "product page" to be `juju.is` rather than
`juju.is/docs/sdk` as the latter is going away (`juju.is` was
recommended by @tmihoc).

(These changes were previously in [a commit in another
PR](https://github.com/tmihoc/operator/commit/54dca52) but got lost when
changes were forced - it seemed cleanest to break them out into a
separate PR).