fix: expose the mocked Juju environment variables via os.environ again (#1573)
Rather than only put `JUJU_VERSION` into `os.environ`, copy the entire
contents of the mock Juju environment variables (and after the run, put
the environment dictionary back to the same state it was originally).

In general, we feel that charms should be using ops to access this
information, rather than going directly to the raw environment variable,
but this was too large a change to make in a refactoring bugfix release,
and should wait for an 8.x release (and also ensuring that there are
reasonable ways to get everything from ops - for example, the event name
is perhaps too tricky to get globally right now).

Fixes #1572

---------

Co-authored-by: James Garner <james.garner@canonical.com>