feat: support starting and stopping Pebble checks, and the checks enabled field (#1560)
Add support for:

* The `startup` field in Pebble checks.
* The `start_checks` and `stop_checks` Pebble API calls.

Harness (and Scenario, mostly via re-using the Harness implementation)
is adjusted to more closely simulate the Changes implementation of
Pebble Checks, so that the 'if the change ID is the empty string, the
check is inactive' behaviour can be simulated.

A subtle bug with notices and check-infos is also fixed: previously the
mocked Pebble would gather all notices and check-infos from all
containers in the state, instead of only those that are in the correct
container.

[Pebble PR](https://github.com/canonical/pebble/pull/560)