docs: add Juju version markers for start/stop checks (#1592)
The version of Pebble that includes the start/stop check functionality
has been [merged into Juju](https://github.com/juju/juju/pull/19037), so
should be in the next Juju release. 3.6.3 is in candidate and 3.6.4 is
edge, so I think that means it'll be in 3.6.4, so set that as the
marker.


[Preview](https://ops--1592.org.readthedocs.build/en/1592/reference/ops.html#ops.Container.start_checks)