feat: expose charm-user in CharmMeta (#1598)
Exposes the `charm-user` field in `CharmMeta`.

The field is added as a string (with type annotations providing the
three possible values) rather than an enum. We have a mixture in the
other meta fields - `RelationRole` is an enum, but `ResourceMeta.type`
is a string and `StorageMeta.type` is a string. It seems like
`charm-user` can go with the simpler form - and it's also easier to add
an enum later than it is to remove one.

Like most of the rest of the meta loading, there's also no validation
(that it's one of the three possible values) - it seems reasonable to
assume that if there's some other value in a production environment
that's a `charmcraft` bug (or there's a new type we don't handle yet)
and in development `charmcraft` will raise the problem when packing.

See also the [charmcraft
docs](https://canonical-charmcraft.readthedocs-hosted.com/en/stable/reference/files/metadata-yaml-file/#charm-user)
(and a [PR to fix it being missing from charmcraft.yaml
reference](https://github.com/canonical/charmcraft/pull/2195)).

Fixes #1396