ci: run tox static environment for charms that have it (#1585)
This PR adds a steps to the `charm-tests` job defined in the
`published-charm-tests` workflow to run the charm's `static` `tox`
environment. The first step checks if the `static` environment exists,
and the second step runs it if it does exist. Additionally, output and
step ids are added to allow these steps to runeven if the unit test step
fails, but not if there was a failure in an earlier step (installing
dependencies, etc.) and not if the job was cancelled. Therefore a given
job may be marked as a failure overall if both or either of the `static`
and `unit` tests fail (if both fail, two failures will be listed for the
job on the summary page for the workflow).

---------

Co-authored-by: Tony Meyer <tony.meyer@gmail.com>