chore: use relative imports (#1606)
Per discussion at daily. I thought this would be a trivial job, but the
Ops imports are a bit of a mess, so it was a bit more work than usual.

As Dima pointed out, we have to be careful in public modules (like
`ops.charm`) not to remove existing names or introduce new names. I
think I've done that right, but please review that carefully. The one
exception is when in `model.py` or similar I've not worried about
removing `ops`, because if you're writing `ops.charm.ops.foo` you
deserve an `AttributeError`, at least. :-)

---------

Co-authored-by: Dima Tisnek <dimaqq@gmail.com>