ci: disable the observability tests requirement in publishing for the 19.1 branch (#1618)
The observability tests need to install ops and ops-scenario, and they
can't do that with the requirements in the branch, because they aren't
available yet.

As a workaround for now, disable the requirement for these tests to pass
in the maintenance branch. This should allow publishing both versions.

For the next release, we'll need to either do this again or implement
the wheel solution or something similar.