chore: bump doc dependencies (#1621)
There's a security alert for Jinja 3.1.5. As far as I can tell, it's not
relevant to our use (which is only in doc generation anyway), but it's
nice to move to the fixed version even if just to silence the warning
systems.

While we're doing this, bump the other doc dependencies (this is just
generating a fresh `requirements.txt` via `uvx --python=3.11 tox -e
docs-deps`).

I haven't read through all of the changes in each of the bumps, but for
the ones that aren't micro bumps, certifi seems safe (and that's
calendar versioning, not semver), sphinx-notfound-page seems safe, the
starlette changes don't seem relevant, and the websockets changes don't
seem relevant to docs (it's less clear for ops itself).