feat: ops.testing.State components are less mutable (#1617)
This PR has two changes to make the `testing.State` less mutable (or
more immuntable as my Kiwi fingers keep typing).

* All of the components that have a dictionary or list attribute now
deepcopy that attribute in `__postinit__`. The type annotations are also
changed to `Mapping` and `Sequence` so static type checking should
detect when charm tests try to alter these (the actual underlying types
are still the ones passed, and default to dict and list).
* In `State.__deepcopy__` we deepcopy all of the items inside the
collections (the relations, the containers, and so forth) - this means
that all of these items are separate in the input and output states.
This also includes the config.

This also fixes a small bug that was found by the new tests:
`State.get_storage` was wrongly comparing the index of the storage
object to the *`str.index` function* (so would never match) rather than
to the index of the `Storage` object in the state. The `| None` type
annotation is also removed, as the index is `int`, not `int | None`, so
it doesn't make sense to pass `None` in as the index.

Fixes #1605