Fixes for merging layers.
When generating the plan, layers should be combined in the order they were added, not alphabetically by label.
when generating the plan, services, checks, and log targets should merge if they have the same name and override set to merge.
Add the threshold to new check infos when adding a layer (oddly missing previously).
After adding a new layer, update any existing check infos to copy the latest values of the checks from the plan."