docs: remove the charm-tech@lists.launchpad.net email address (#1632)
We've had issues with the reliability of the `charm-tech` launchpad
email address, and we don't ever actually get any contact via it.

This PR therefore drops the `charm-tech@lists.launchpad.net` email
address from `pyproject.toml` (and therefore PyPI) for both `ops` and
`ops-scenario`. Both projects have links to the GitHub repositories
where issues and PRs and security reports can be created, and it's also
pretty clear from both repositories who is actively working on the
projects and so if someone needs to be privately contacted, they can
always go via the GitHub profile.

Note that our CoC has an address `charmcrafters@lists.launchpad.net` --
I'm assuming that still works and is the correct one to have (a private
contact method in the CoC does make sense).