fix: assorted fixes for Pebble layer merging in Harness and Scenario (#1627)
* When merging `pebble.Check`s, an unset `startup` should not override a
set `startup`. (Once set you can't go back to unset - you need to
explicitly use enabled).
* When rendering services, checks, and log targets from a series of
layers, Harness and Scenario were combining after sorting the items by
name. Although Pebble uses filename sorting when reading initial layers
from a directory, future layers are combined in the order they are added
via the API. For Harness and Scenario, we need to use the order in which
they were added - if the caller wants to simulate loading from the
filesystem, they need to do that by calling `add_layer` in that order.
The plan itself still sorts the services, checks, and log targets.
* When rendering services, checks, and log targets, if there are
multiple services/checks/log-targets with the same name, and `override`
is set to `merge`, the service/check/log-target needs to merge rather
than replace. This was already done when adding layers with the same
name (in `add_layer`) but not when there were differently named layers.
* The Scenario `Container.plan` property was ignoring any defined checks
and log targets. Those are now included.
* The Scenario `Container` service, check, and log target rendering also
needed the (no) sorting and merging fixes from above.
* When adding a layer in Harness or Scenario, update any existing
`CheckInfo`s to match changes to the plan - in real Pebble the startup,
level, and threshold fields are always pulled from the plan but we bake
them in when testing, so need to make sure they continually track what's
in the combined plan.

Also adds a `__repr__` for `pebble.Plan` - this helped while debugging
issues, and it seems reasonable to keep it.

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>